Move the extensions into the patches directory
This way the patch code can stay together
